+++ library/src/methods/getDefaults/index.ts
@@ -1,0 +1,3 @@
+export * from './getDefaults.ts';
+export * from './getDefaultsAsync.ts';
+export * from './types.ts';